requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Django==4.2.7
django-rest-knox==4.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
dynaconf==3.2.4
hvac==2.0.0
idna==3.6
loguru==0.7.2
pycparser==2.21
pytz==2023.3.post1
requests==2.31.0
shortuuid==1.0.11
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
validators==0.22.0
win32-setctime==1.1.0
